pyproject.toml
[project]
name = "rift"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gdm-concordia",
    "httpx[socks]>=0.28.1",
    "networkx",
    "python-dotenv",
    "sentence-transformers",
]
